Initial work to set up dynamic recipe details page.
