+++ pages/recipe/[slug].js
@@ -1,29 +1,56 @@
-import React, {Component} from 'react'
+import React, { Component } from 'react';
+import RecipeDetail from '../../components/RecipeDetail';
+
 
 class Post extends Component {
-    static async getInitialProps({query}){
-        const {slug} = query;
-        const recipePost = await import(`../../content/recipes/${slug}.md`).catch(error => null);
+  static async getInitialProps({ query }) {
+    const { slug } = query;
+    const recipePost = await import(`../../content/recipes/${slug}.md`).catch(
+      (error) => null
+    );
 
-        return {recipePost};
-    }
+    return { recipePost };
+  }
 
-    render() {
-        if(!this.props.recipePost) return <div>not found</div>;
+  render() {
+    if (!this.props.recipePost)
+      // Create a error page
+      return <div>not found</div>;
 
-        const {
-            html,
-            attributes
-        } = this.props.recipePost.default;
+    const { html, attributes } = this.props.recipePost.default;
 
-        console.log(attributes)
+    console.log(attributes)
+    
+    let {title,
+    author,
+    summary,
+    dairyFree,
+    servings,
+    prepTime,
+    cookTime,
+    featuredImage,
+    ingredients,
+    instructions,
+    videoTitle,
+    videoURL} = attributes;
 
-        return(
-            <>
-                
-            </>
-        );
-    }
+    return (
+      <>
+        <RecipeDetail title={attributes.title},
+    author={attributes.author},
+    summary={attributes.summary},
+    dairyFree={attributes.dairyFree},
+    servings={attributes.servings},
+    prepTime={attributes.prepTime},
+    cookTime={attributes.cookTime},
+    featuredImage,
+    ingredients,
+    instructions,
+    videoTitle,
+    videoURL,/>
+      </>
+    );
+  }
 }
 
 export default Post;
